[#233] move data into /data
