+++ components/CalendarEvents.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import moment from "moment";
 import { Calendar, momentLocalizer } from "react-big-calendar";
 import "react-big-calendar/lib/css/react-big-calendar.css";
-import { Events } from "./Events.js";
+import { Events } from "./data/Events.js";
 import { Row, Col } from "react-bootstrap";
 import CustomToolbar from "./CustomToolbar.js";
 import CustomEvent from "./CustomEvent.js";
