+++ components/Faq.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Accordion from "react-bootstrap/Accordion";
-import { FaqInfoArray } from "./FaqInfoArray";
+import { FaqInfoArray } from "./data/FaqInfoArray";
 
 let colorCount = 0;
